ASAP-1607 Add showImpactAndCategory to AvailableActions
