+++ packages/react-components/src/organisms/__tests__/ResearchOutputFormSharingCard.test.tsx
@@ -1,23 +1,25 @@
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
+import { researchTagSubtypeResponse } from '@asap-hub/fixtures';
+import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { ComponentProps } from 'react';
-import { MultiSelectOptionsType } from '../../atoms';
+import { editorRef, MultiSelectOptionsType } from '../../atoms';
 import { OptionsType } from '../../select';
-import { noop } from '../../utils';
-
+import { mockActErrorsInConsole } from '../../test-utils';
 import ResearchOutputFormSharingCard from '../ResearchOutputFormSharingCard';
 
 const defaultProps: ComponentProps<typeof ResearchOutputFormSharingCard> = {
+  showImpactAndCategory: false,
   isFormSubmitted: false,
-  documentType: 'Bioinformatics',
+  isCreatingNewVersion: false,
+  isSaving: false,
   link: '',
   title: '',
+  type: '',
   descriptionMD: '',
   shortDescription: '',
   changelog: '',
   subtype: '',
   displayChangelog: true,
-  isSaving: false,
   researchTags: [],
   typeOptions: [],
   getShortDescriptionFromDescription: jest.fn(),
@@ -34,11 +36,240 @@ const defaultProps: ComponentProps<typeof ResearchOutputFormSharingCard> = {
   getCategorySuggestions: jest.fn().mockResolvedValue([]),
   categories: undefined,
   onChangeCategories: jest.fn(),
-  isCreatingNewVersion: false,
 };
 
+let consoleMock: ReturnType<typeof mockActErrorsInConsole>;
 beforeEach(() => {
-  jest.spyOn(console, 'error').mockImplementation();
+  consoleMock = mockActErrorsInConsole();
+});
+afterEach(() => {
+  consoleMock.mockRestore();
+});
+
+describe('base sharing fields', () => {
+  it('renders the card with provided values', () => {
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        descriptionMD="description text"
+        link="http://example.com"
+        title="title"
+        type={'Preprint'}
+        typeOptions={['Preprint', '3D Printing']}
+      />,
+    );
+    expect(screen.getAllByText('description text')[0]).toBeVisible();
+    expect(screen.getByDisplayValue('http://example.com')).toBeVisible();
+    expect(screen.getByDisplayValue('title')).toBeVisible();
+    expect(screen.getByText('Preprint')).toBeVisible();
+  });
+
+  it.each`
+    title      | label       | error
+    ${'Url'}   | ${/URL/i}   | ${'Please enter a valid URL, starting with http://'}
+    ${'Title'} | ${/title/i} | ${'Please enter a title'}
+    ${'Type'}  | ${/type/i}  | ${'Please choose a type'}
+  `(
+    'shows error message for missing value $title',
+    async ({ label, error }) => {
+      const { findByText } = render(
+        <ResearchOutputFormSharingCard
+          {...defaultProps}
+          urlRequired
+          typeOptions={['3D Printing']}
+        />,
+      );
+      const input = screen.getByLabelText(label);
+      await userEvent.click(input);
+      await userEvent.tab();
+      expect(await findByText(error)).toBeVisible();
+    },
+  );
+
+  it('does not require an url', async () => {
+    render(
+      <ResearchOutputFormSharingCard {...defaultProps} urlRequired={false} />,
+    );
+    expect(
+      await screen.findByText(
+        (content, node) =>
+          (content === 'URL' &&
+            node?.nextSibling?.textContent?.includes('(optional)')) ??
+          false,
+      ),
+    ).toBeVisible();
+
+    const input = screen.getByLabelText(/URL/i);
+    fireEvent.focusOut(input);
+    await expect(
+      screen.findByText('Please enter a valid URL, starting with http://'),
+    ).rejects.toThrow();
+  });
+
+  it.each`
+    field      | label       | prop
+    ${'Url'}   | ${/URL/i}   | ${'onChangeLink'}
+    ${'Title'} | ${/title/i} | ${'onChangeTitle'}
+  `('triggers an onchange event for $field', async ({ label, prop }) => {
+    const onChangeFn = jest.fn();
+    render(
+      <ResearchOutputFormSharingCard
+        {...{ ...defaultProps, [prop]: onChangeFn }}
+      />,
+    );
+    const input = screen.getByLabelText(label);
+    fireEvent.change(input, { target: { value: 'test' } });
+    expect(onChangeFn).toHaveBeenLastCalledWith('test');
+  });
+
+  it('triggers an onchange event for Description', async () => {
+    const onChangeFn = jest.fn();
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        onChangeDescription={onChangeFn}
+      />,
+    );
+    await waitFor(() => expect(editorRef.current).not.toBeNull());
+
+    editorRef.current?.focus();
+    const input = screen.getByTestId('editor');
+
+    await userEvent.click(input);
+    await userEvent.tab();
+    fireEvent.input(input, { data: 'test' });
+
+    await waitFor(() => {
+      expect(onChangeFn).toHaveBeenLastCalledWith('test');
+    });
+  });
+
+  it('triggers an onchange event for Short Description', async () => {
+    const onChangeFn = jest.fn();
+    render(
+      <ResearchOutputFormSharingCard
+        {...{ ...defaultProps, onChangeShortDescription: onChangeFn }}
+      />,
+    );
+    const input = screen.getByRole('textbox', { name: /short description/i });
+    fireEvent.change(input, { target: { value: 'test' } });
+    await waitFor(() => {
+      expect(onChangeFn).toHaveBeenCalledWith('test');
+    });
+  });
+
+  it('triggers an on change for type', async () => {
+    const onChangeFn = jest.fn();
+
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        typeOptions={['Preprint', '3D Printing', 'ASAP subgroup meeting']}
+        onChangeType={onChangeFn}
+      />,
+    );
+
+    const type = screen.getByLabelText(/type/i);
+    await userEvent.type(type, 'Preprint');
+    await userEvent.type(type, '{Enter}');
+
+    expect(onChangeFn).toHaveBeenCalledWith('Preprint');
+  });
+
+  it('triggers an on change for subtype', async () => {
+    const onChangeFn = jest.fn();
+
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        researchTags={[researchTagSubtypeResponse]}
+        onChangeSubtype={onChangeFn}
+      />,
+    );
+
+    const type = screen.getByLabelText(/subtype/i);
+    await userEvent.type(type, 'Metabolite');
+    await userEvent.type(type, '{Enter}');
+
+    expect(onChangeFn).toHaveBeenCalledWith('Metabolite');
+  });
+
+  it('shows the custom no options message for type', async () => {
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        typeOptions={['ASAP annual meeting', '3D Printing']}
+      />,
+    );
+
+    await userEvent.type(screen.getByLabelText(/type/i), 'asdflkjasdflkj');
+
+    expect(
+      screen.getByText('Sorry, no types match asdflkjasdflkj'),
+    ).toBeVisible();
+  });
+
+  it('displays server side validation error for link and calls clears function when changed', async () => {
+    const mockClearError = jest.fn();
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        link="http://example.com"
+        serverValidationErrors={[
+          {
+            instancePath: '/link',
+            keyword: '',
+            params: {},
+            schemaPath: '',
+          },
+        ]}
+        clearServerValidationError={mockClearError}
+      />,
+    );
+    expect(
+      screen.getByText(
+        'A Research Output with this URL already exists. Please enter a different URL.',
+      ),
+    ).toBeVisible();
+
+    const input = screen.getByLabelText(/URL/i);
+    await userEvent.type(input, 'a');
+
+    await waitFor(() => {
+      expect(mockClearError).toHaveBeenCalledWith('/link');
+    });
+  });
+
+  it('displays server side validation error for title and calls clears function when changed', async () => {
+    const mockClearError = jest.fn();
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        title="Example"
+        serverValidationErrors={[
+          {
+            instancePath: '/title',
+            keyword: '',
+            params: {},
+            schemaPath: '',
+          },
+        ]}
+        clearServerValidationError={mockClearError}
+      />,
+    );
+    expect(
+      screen.getByText(
+        'A Research Output with this title already exists. Please check if this is repeated and choose a different title.',
+      ),
+    ).toBeVisible();
+
+    const input = screen.getByLabelText(/title/i);
+    await userEvent.type(input, 'a');
+
+    await waitFor(() => {
+      expect(mockClearError).toHaveBeenCalledWith('/title');
+    });
+  });
 });
 
 describe('Changelog validation', () => {
@@ -86,270 +317,274 @@ describe('Changelog validation', () => {
   });
 });
 
-it('renders the impact and categories', async () => {
-  const loadOptions = jest.fn();
-  loadOptions.mockResolvedValue([]);
-  render(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      getImpactSuggestions={loadOptions}
-      getCategorySuggestions={loadOptions}
-    />,
-  );
+describe('impact and category', () => {
+  it('renders the impact and categories', async () => {
+    const loadOptions = jest.fn();
+    loadOptions.mockResolvedValue([]);
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        getImpactSuggestions={loadOptions}
+        getCategorySuggestions={loadOptions}
+      />,
+    );
 
-  const impactInput = screen.getByRole('combobox', { name: /impact/i });
-  expect(impactInput).toBeInTheDocument();
+    const impactInput = screen.getByRole('combobox', { name: /impact/i });
+    expect(impactInput).toBeInTheDocument();
 
-  const categoryInput = screen.getByRole('combobox', { name: /category/i });
-  expect(categoryInput).toBeInTheDocument();
+    const categoryInput = screen.getByRole('combobox', { name: /category/i });
+    expect(categoryInput).toBeInTheDocument();
 
-  await waitFor(() => {
-    expect(screen.queryByText(/loading/i)).not.toBeInTheDocument();
-  });
+    fireEvent.change(categoryInput, { target: { value: 'Category' } });
 
-  fireEvent.change(categoryInput, { target: { value: 'Category' } });
+    await waitFor(() => {
+      expect(screen.queryByText(/loading/i)).not.toBeInTheDocument();
+    });
 
-  await waitFor(() => {
-    expect(screen.queryByText(/loading/i)).not.toBeInTheDocument();
-  });
-
-  expect(
-    screen.getByText(/no category options match Category/i),
-  ).toBeInTheDocument();
-});
-
-it('renders category input and does not throw when getCategorySuggestions is noop', async () => {
-  const { getCategorySuggestions, ...propsWithoutCategorySuggestions } =
-    defaultProps;
-  render(
-    <ResearchOutputFormSharingCard
-      {...propsWithoutCategorySuggestions}
-      documentType="Article"
-      getCategorySuggestions={undefined}
-    />,
-  );
-
-  const categoryInput = screen.getByRole('combobox', { name: /category/i });
-  expect(categoryInput).toBeInTheDocument();
-
-  fireEvent.change(categoryInput, { target: { value: 'Test' } });
-
-  await waitFor(() => {
     expect(
-      screen.queryByText(/no category options match/i),
+      screen.getByText(/no category options match Category/i),
+    ).toBeInTheDocument();
+  });
+
+  it('renders the category field with default callbacks without crashing', async () => {
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        getCategorySuggestions={undefined}
+      />,
+    );
+
+    const categoryInput = screen.getByRole('combobox', { name: /category/i });
+    expect(categoryInput).toBeInTheDocument();
+
+    fireEvent.change(categoryInput, { target: { value: 'Test' } });
+
+    await waitFor(() => {
+      expect(
+        screen.queryByText(/no category options match/i),
+      ).not.toBeInTheDocument();
+    });
+  });
+
+  it('calls onChangeCategories when a category is selected', async () => {
+    const onChangeCategories = jest.fn();
+    const getCategorySuggestions = jest
+      .fn()
+      .mockResolvedValue([{ label: 'Cat 1', value: 'cat1' }]);
+
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        onChangeCategories={onChangeCategories}
+        getCategorySuggestions={getCategorySuggestions}
+      />,
+    );
+
+    const categoryInput = screen.getByRole('combobox', { name: /category/i });
+    fireEvent.change(categoryInput, { target: { value: 'Cat' } });
+
+    await waitFor(() =>
+      expect(screen.queryByText(/loading/i)).not.toBeInTheDocument(),
+    );
+
+    fireEvent.keyDown(categoryInput, { key: 'ArrowDown' });
+    fireEvent.keyDown(categoryInput, { key: 'Enter' });
+
+    expect(onChangeCategories).toHaveBeenCalled();
+  });
+
+  it('shows validation message when more than two categories are selected', async () => {
+    const getCategorySuggestions = jest
+      .fn()
+      .mockResolvedValue([{ label: 'Cat 3', value: 'cat3' }]);
+
+    const onChangeCategories = (
+      newCategories: OptionsType<MultiSelectOptionsType>,
+    ) => {
+      rerender(
+        <ResearchOutputFormSharingCard
+          {...defaultProps}
+          showImpactAndCategory={true}
+          onChangeCategories={onChangeCategories}
+          getCategorySuggestions={getCategorySuggestions}
+          categories={newCategories}
+        />,
+      );
+    };
+
+    const { rerender } = render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        onChangeCategories={onChangeCategories}
+        getCategorySuggestions={getCategorySuggestions}
+        categories={[
+          { label: 'Cat 1', value: 'cat1' },
+          { label: 'Cat 2', value: 'cat2' },
+        ]}
+      />,
+    );
+
+    const categoryInput = screen.getByRole('combobox', { name: /category/i });
+    expect(categoryInput).toBeInTheDocument();
+
+    fireEvent.change(categoryInput, { target: { value: 'Cat' } });
+
+    await waitFor(() =>
+      expect(screen.queryByText(/loading/i)).not.toBeInTheDocument(),
+    );
+
+    fireEvent.keyDown(categoryInput, { key: 'ArrowDown' });
+    fireEvent.keyDown(categoryInput, { key: 'Enter' });
+
+    fireEvent.blur(categoryInput);
+
+    expect(
+      await screen.findByText('You can select up to two categories only.'),
+    ).toBeInTheDocument();
+
+    fireEvent.click(screen.getAllByTitle(/close/i)[0]!);
+    fireEvent.blur(categoryInput);
+
+    expect(
+      screen.queryByText('You can select up to two categories only.'),
     ).not.toBeInTheDocument();
   });
-});
 
-it('renders category input and does not throw when onChangeCategories is noop', async () => {
-  const { onChangeCategories, ...propsWithoutOnChangeCategories } =
-    defaultProps;
-  render(
-    <ResearchOutputFormSharingCard
-      {...propsWithoutOnChangeCategories}
-      documentType="Article"
-      onChangeCategories={noop}
-    />,
-  );
+  it('disables impact and category fields when isCreatingNewVersion is true', () => {
+    const loadOptions = jest.fn();
+    loadOptions.mockResolvedValue([]);
 
-  const categoryInput = screen.getByRole('combobox', { name: /category/i });
-  expect(categoryInput).toBeInTheDocument();
+    const { rerender } = render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        getImpactSuggestions={loadOptions}
+        getCategorySuggestions={loadOptions}
+        isCreatingNewVersion={true}
+      />,
+    );
+    expect(screen.getByLabelText(/impact\(required\)/i)).toBeDisabled();
+    expect(screen.getByLabelText(/category/i)).toBeDisabled();
 
-  fireEvent.change(categoryInput, { target: { value: 'Test' } });
-
-  await waitFor(() => {
-    expect(
-      screen.queryByText(/no category options match/i),
-    ).not.toBeInTheDocument();
-  });
-});
-
-it('shows validation message when more than two categories are selected', async () => {
-  const getCategorySuggestions = jest
-    .fn()
-    .mockResolvedValue([{ label: 'Cat 3', value: 'cat3' }]);
-
-  const onChangeCategories = (
-    newCategories: OptionsType<MultiSelectOptionsType>,
-  ) => {
     rerender(
       <ResearchOutputFormSharingCard
         {...defaultProps}
-        documentType="Article"
-        onChangeCategories={onChangeCategories}
-        getCategorySuggestions={getCategorySuggestions}
-        categories={newCategories}
+        showImpactAndCategory={true}
+        getImpactSuggestions={loadOptions}
+        getCategorySuggestions={loadOptions}
+        isSaving={false}
+        isCreatingNewVersion={false}
       />,
     );
-  };
-
-  const { rerender } = render(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      onChangeCategories={onChangeCategories}
-      getCategorySuggestions={getCategorySuggestions}
-      categories={[
-        { label: 'Cat 1', value: 'cat1' },
-        { label: 'Cat 2', value: 'cat2' },
-      ]}
-    />,
-  );
-
-  const categoryInput = screen.getByRole('combobox', { name: /category/i });
-  expect(categoryInput).toBeInTheDocument();
-
-  fireEvent.change(categoryInput, { target: { value: 'Cat' } });
-
-  await waitFor(() =>
-    expect(screen.queryByText(/loading/i)).not.toBeInTheDocument(),
-  );
-
-  fireEvent.keyDown(categoryInput, { key: 'ArrowDown' });
-  fireEvent.keyDown(categoryInput, { key: 'Enter' });
-
-  fireEvent.blur(categoryInput);
-
-  expect(
-    await screen.findByText('You can select up to two categories only.'),
-  ).toBeInTheDocument();
-
-  fireEvent.click(screen.getAllByTitle(/close/i)[0]!);
-  fireEvent.blur(categoryInput);
-
-  expect(
-    screen.queryByText('You can select up to two categories only.'),
-  ).not.toBeInTheDocument();
-});
-
-it('disables impact and category fields when isCreatingNewVersion is true', () => {
-  const loadOptions = jest.fn();
-  loadOptions.mockResolvedValue([]);
-
-  const { rerender } = render(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      getImpactSuggestions={loadOptions}
-      getCategorySuggestions={loadOptions}
-      isCreatingNewVersion={true}
-    />,
-  );
-  expect(screen.getByLabelText(/impact\(required\)/i)).toBeDisabled();
-  expect(screen.getByLabelText(/category/i)).toBeDisabled();
-
-  rerender(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      getImpactSuggestions={loadOptions}
-      getCategorySuggestions={loadOptions}
-      isSaving={false}
-      isCreatingNewVersion={false}
-    />,
-  );
-  expect(screen.getByLabelText(/impact\(required\)/i)).toBeEnabled();
-  expect(screen.getByLabelText(/category/i)).toBeEnabled();
-});
-
-it('calls onChangeImpact with the correct option when an impact is selected', async () => {
-  const impactOptions = [
-    { label: 'High Impact', value: 'high' },
-    { label: 'Low Impact', value: 'low' },
-  ];
-  const getImpactSuggestions = jest.fn().mockResolvedValue(impactOptions);
-  const onChangeImpact = jest.fn();
-
-  render(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      impact={undefined}
-      onChangeImpact={onChangeImpact}
-      getImpactSuggestions={getImpactSuggestions}
-    />,
-  );
-
-  await waitFor(() => {
-    expect(getImpactSuggestions).toHaveBeenCalled();
+    expect(screen.getByLabelText(/impact\(required\)/i)).toBeEnabled();
+    expect(screen.getByLabelText(/category/i)).toBeEnabled();
   });
 
-  const impactField = screen.getByLabelText(/impact\(required\)/i);
-  await userEvent.click(impactField);
+  it('calls onChangeImpact with the correct option when an impact is selected', async () => {
+    const impactOptions = [
+      { label: 'High Impact', value: 'high' },
+      { label: 'Low Impact', value: 'low' },
+    ];
+    const getImpactSuggestions = jest.fn().mockResolvedValue(impactOptions);
+    const onChangeImpact = jest.fn();
 
-  await waitFor(() => {
-    expect(screen.getByText(/low impact/i)).toBeInTheDocument();
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        impact={undefined}
+        onChangeImpact={onChangeImpact}
+        getImpactSuggestions={getImpactSuggestions}
+      />,
+    );
+
+    await waitFor(() => {
+      expect(getImpactSuggestions).toHaveBeenCalled();
+    });
+
+    const impactField = screen.getByLabelText(/impact\(required\)/i);
+    await userEvent.click(impactField);
+
+    await waitFor(() => {
+      expect(screen.getByText(/low impact/i)).toBeInTheDocument();
+    });
+
+    await userEvent.click(screen.getByText('Low Impact'));
+
+    expect(onChangeImpact).toHaveBeenCalledWith(
+      expect.objectContaining({ label: 'Low Impact', value: 'low' }),
+    );
   });
 
-  await userEvent.click(screen.getByText('Low Impact'));
+  it('shows validation message when impact is not selected', async () => {
+    const getImpactSuggestions = jest.fn().mockResolvedValue([
+      { label: 'High Impact', value: 'high' },
+      { label: 'Low Impact', value: 'low' },
+    ]);
 
-  expect(onChangeImpact).toHaveBeenCalledWith(
-    expect.objectContaining({ label: 'Low Impact', value: 'low' }),
-  );
-});
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        isFormSubmitted={true}
+        impact={undefined}
+        getImpactSuggestions={getImpactSuggestions}
+      />,
+    );
 
-it('shows validation message when impact is not selected', async () => {
-  const getImpactSuggestions = jest.fn().mockResolvedValue([
-    { label: 'High Impact', value: 'high' },
-    { label: 'Low Impact', value: 'low' },
-  ]);
+    await waitFor(() => {
+      expect(
+        screen.getByText('Please add at least one impact.'),
+      ).toBeInTheDocument();
+    });
+  });
 
-  render(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      isFormSubmitted={true}
-      impact={undefined}
-      getImpactSuggestions={getImpactSuggestions}
-    />,
-  );
+  it('does not show impact validation message when impact is selected', async () => {
+    const getImpactSuggestions = jest.fn().mockResolvedValue([
+      { label: 'High Impact', value: 'high' },
+      { label: 'Low Impact', value: 'low' },
+    ]);
 
-  await waitFor(() => {
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+        isFormSubmitted={true}
+        impact={{ label: 'High Impact', value: 'high' }}
+        getImpactSuggestions={getImpactSuggestions}
+      />,
+    );
+
+    await waitFor(() => {
+      expect(
+        screen.queryByText('Please add at least one impact.'),
+      ).not.toBeInTheDocument();
+    });
+  });
+
+  it('shows validation message when lay impact statement exceeds 100 characters', async () => {
+    render(
+      <ResearchOutputFormSharingCard
+        {...defaultProps}
+        showImpactAndCategory={true}
+      />,
+    );
+
+    const longText = 'a'.repeat(101);
+    const layImpactInput = screen.getByRole('textbox', {
+      name: /lay impact statement/i,
+    });
+    fireEvent.change(layImpactInput, { target: { value: longText } });
+    fireEvent.blur(layImpactInput);
+
     expect(
-      screen.getByText('Please add at least one impact.'),
+      await screen.findByText(
+        'The lay impact statement exceeds the character limit. Please limit it to 100 characters.',
+      ),
     ).toBeInTheDocument();
   });
-});
-
-it('does not show impact validation message when impact is selected', async () => {
-  const getImpactSuggestions = jest.fn().mockResolvedValue([
-    { label: 'High Impact', value: 'high' },
-    { label: 'Low Impact', value: 'low' },
-  ]);
-
-  render(
-    <ResearchOutputFormSharingCard
-      {...defaultProps}
-      documentType="Article"
-      isFormSubmitted={true}
-      impact={{ label: 'High Impact', value: 'high' }}
-      getImpactSuggestions={getImpactSuggestions}
-    />,
-  );
-
-  await waitFor(() => {
-    expect(
-      screen.queryByText('Please add at least one impact.'),
-    ).not.toBeInTheDocument();
-  });
-});
-
-it('shows validation message when lay impact statement exceeds 100 characters', async () => {
-  render(
-    <ResearchOutputFormSharingCard {...defaultProps} documentType="Article" />,
-  );
-
-  const longText = 'a'.repeat(101);
-  const layImpactInput = screen.getByRole('textbox', {
-    name: /lay impact statement/i,
-  });
-  fireEvent.change(layImpactInput, { target: { value: longText } });
-  fireEvent.blur(layImpactInput);
-
-  expect(
-    await screen.findByText(
-      'The lay impact statement exceeds the character limit. Please limit it to 100 characters.',
-    ),
-  ).toBeInTheDocument();
 });
