+++ packages/react-components/src/templates/ResearchOutputForm.tsx
@@ -518,10 +518,10 @@ const ResearchOutputForm: React.FC<ResearchOutputFormProps> = ({
               )}
               <div css={contentStyles} data-flow-id={flowId}>
                 <ResearchOutputFormSharingCard
+                  showImpactAndCategory={availableActions.showImpactAndCategory}
                   isFormSubmitted={isFormSubmitted}
                   isCreatingNewVersion={versionAction === 'create'}
                   displayChangelog={displayChangelog}
-                  documentType={documentType}
                   serverValidationErrors={serverValidationErrors}
                   clearServerValidationError={clearServerValidationError}
                   isSaving={isSaving}
