+++ packages/react-components/src/templates/__tests__/ResearchOutputForm/ResearchOutputForm.formButtons.test.tsx
@@ -104,6 +104,7 @@ describe('form buttons', () => {
             availableActions={resolveResearchOutputAvailableActions({
               flowId,
               permissions,
+              documentType,
             })}
             permissions={permissions}
           />
